Prototype the first inbox async worker
A worker pool selects the worker using the hash_worker strategy, on the
inbox entry key. This way, a single worker will be responsible for
dealing with all the events for this key, being therefore able to ensure
linearisation, and not only but also to aggregate them. That is, two
incoming messages, instead of doing two DB operations with two +1
increments, can instead aggregate the increments and do a single DB
operation at the end.

Then, all the prepared operations can be flushed to the DB worker in a
single round-trip, to reduce, precisely, trips back and forth.