[WIP][RDBMS] Ensure ordering of messages on the DB layer
Prototype a new parameter to the rdbms prepared queries, that would
allow to provide an extra clause to the statement, which introduces a
`WHERE` clause to postgres' ON CONFLICT. This way, we try to ensure that
we don't update inbox with messages older than the currently inserted,
so that the inbox never recedes.

This way we increase the consistency of inbox entries being aggregated
across nodes in a cluster, while we don't have a remedy for incorrect
inbox resets (like when sending a marker). This is not too bad,
resending messages is a very undesirable event, but just resending
markers is a common behaviour in client devices, and it is then
eventually consistent.